Rename io.{h,cc} to pio.{h,cc} to avoid naming conflict with Windows-provided io.h.
