+++ ecp5/bitstream.cc
@@ -26,7 +26,7 @@
 #include <streambuf>
 
 #include "config.h"
-#include "io.h"
+#include "pio.h"
 #include "log.h"
 #include "util.h"
